Move the implementation of compute_scale() to utils.rs.
